+++ R/poisson.R
@@ -37,7 +37,7 @@ sufficientT.trunc_poisson <- function(y) {
 }
 
 #' @export
-natural2parameters.trunc_poisson <- function(eta) {
+natural2parameters.trunc_poisson <- function(eta, ...) {
   # eta: The natural parameters in a Poisson distribution
   # returns (mean,sigma)
   if (length(eta) != 1) stop("Eta must be one single number")
@@ -47,7 +47,7 @@ natural2parameters.trunc_poisson <- function(eta) {
 }
 
 #' @export
-parameters2natural.trunc_poisson <- function(parms) {
+parameters2natural.trunc_poisson <- function(parms, ...) {
   # parms: The parameter lambda in a Poisson distribution
   # returns the natural parameters
   eta <- prepEta(log(parms), class(parms))